implemented method for generating refs, long name, shortname, commit, category, lcstatus, fingerpirnt

+++ code/src/main/java/amos/specitemdatabase/filegenerator/SpecItemProvider.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public interface SpecItemProvider {
 
 
-    List<Map<String, String>> generateSpecItems(boolean complete, short numberOfSpecItems);
+    List<Map<String, String>> generateSpecItems(boolean complete, int numberOfSpecItems);
 
 
 }
